added colorButtonView to CommonUI and changed button in LoginViewController
